nmap-requirements.txt
# Core dependencies
python-nmap>=0.7.1
tqdm>=4.65.0

# Testing and development
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development tools (optional)
flake8>=6.0.0
black>=23.3.0
